+++ packages/ui-components/src/postgrest-infinite-scroll.ts
@@ -8,9 +8,12 @@ import { MenuItem, Spinner, InputGroup } from "@blueprintjs/core";
 interface PostgRESTInfiniteScrollProps extends InfiniteScrollProps<any> {
   id_key: string;
   limit: number;
+  extraParams?: Record<string, any>;
   ascending?: boolean;
   filterable?: boolean;
   order_key?: string;
+  key?: string;
+  toggles: any;
   SearchBarComponent?: React.ComponentType<{
     onChange: (value: string) => void;
   }>;
@@ -43,6 +46,8 @@ export function PostgRESTInfiniteScrollView(
     order_key = undefined,
     SearchBarComponent,
     MultiSelectComponent,
+    extraParams = {},
+    key,
     ...rest
   } = props;
 
@@ -63,6 +68,10 @@ export function PostgRESTInfiniteScrollView(
   const notOperator2 = ascending ? `lt` : `gt`;
   const id = ascending ? 0 : maxId;
   const notId = ascending ? maxId : 0;
+  const newInitialItems =
+    selectedItems.length === 0 && filterValue === "" ? initialItems : undefined;
+
+  console.log("initialItems", initialItems);
 
   const orParam = `(${order_key}.${operator2}.${id},and(${order_key}.eq.${id},${id_key}.${notOperator2}.${notId}))`;
 
@@ -75,8 +84,9 @@ export function PostgRESTInfiniteScrollView(
 
   const defaultParams = useMemo(() => {
     return {
+      ...extraParams,
       [id_key]: !order_key
-        ? operator2 + `.${initialItems?.[0]?.[id_key] ?? id}`
+        ? operator2 + `.${newInitialItems?.[0]?.[id_key] ?? id}`
         : undefined,
       order: order_key
         ? `${order_key}.${operator1},${id_key}.${notOperator1}`
@@ -96,7 +106,7 @@ export function PostgRESTInfiniteScrollView(
     filterValue,
     orParam,
     searchItemsParam,
-    initialItems,
+    newInitialItems,
     id_key,
     order_key,
     limit,
@@ -186,6 +196,9 @@ export function PostgRESTInfiniteScrollView(
     });
   };
 
+  const newKey =
+    key || `${filterValue}-${selectedItems.join(",")}-${props.toString()}`;
+
   return h("div.postgrest-infinite-scroll", [
     h.if(filterable)("div.search-bar", [
       h(SearchBarToUse, {
@@ -211,9 +224,9 @@ export function PostgRESTInfiniteScrollView(
       route,
       getNextParams: getNextParams ?? defaultGetNextParams,
       params: params ?? defaultParams,
-      initialItems,
+      initialItems: newInitialItems,
       hasMore: hasMore ?? defaultHasMore,
-      key: filterValue + selectedItems.join(","),
+      key: newKey,
     }),
   ]);
 }