+++ packages/ui-components/CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## [4.3.5] - 2025-07-10
+
+- Handle extra params and intitial items
+- Better state management
+
 ## [4.3.5] - 2025-07-09
 
 - Fix import issue